feat: add the cost estimate integration to qhub
