--- qhub/_version.py
@@ -1,5 +1,0 @@
-# coding: utf-8
-# file generated by setuptools_scm
-# don't change, don't track in version control
-version = "0.5.0.dev18+g7ca7f2f6.d20220621155601"
-version_tuple = (0, 5, 0, "dev18", "g7ca7f2f6.d20220621155601")
